requirements.txt
pillow
face_recognition